WIP: nodejs glue — re-pin to FINAL PIN 857783e6 (status strings + URL/PUBKEY)
Bump the vendored libsession submodule 73e4d1e6 -> 857783e6 (the FINAL PIN) and
adapt to the two changes since 73e4d1e6:

- get-details `status` / `user_status` are now opaque string codes (the
  SESSION_PRO_BACKEND_PAYMENT_STATUS / _USER_PRO_STATUS enums were deleted):
  parsePaymentDetailsResponse now emits them as pass-through strings
  (item status: unredeemed/redeemed/expired/revoked; userStatus:
  never/active/expired). error_report stays a numeric enum. pro.d.ts retyped.
- Expose the new libsession-owned backend identity so desktop stops
  hand-carrying its own: LIBSESSION_PRO_BACKEND_URL (session::pro_backend::URL,
  the overridable prod default) + LIBSESSION_PRO_BACKEND_PUBKEY_HEX
  (hex of session::pro_backend::PUBKEY) on the CONSTANTS wrapper; declared in
  shared.d.ts. Desktop will read these and derive X25519 on the fly.

+++ include/pro/pro.hpp
@@ -347,7 +347,8 @@ class ProWrapper : public Napi::ObjectWrap<ProWrapper> {
             auto obj = Napi::Object::New(env);
             obj["status"] = toJs(env, resp.status);
             obj["errors"] = errorsToJs(env, resp.errors);
-            obj["userStatus"] = toJs(env, static_cast<uint32_t>(resp.user_status));
+            // user_status is now an opaque string code (never/active/expired; unknowns pass through)
+            obj["userStatus"] = toJs(env, resp.user_status);
             obj["errorReport"] = toJs(env, static_cast<uint32_t>(resp.error_report));
             obj["autoRenewing"] = toJs(env, resp.auto_renewing);
             obj["expiryMs"] = toJs(env, resp.expiry_unix_ts.time_since_epoch().count() * 1000);
@@ -361,7 +362,8 @@ class ProWrapper : public Napi::ObjectWrap<ProWrapper> {
             for (size_t i = 0; i < resp.items.size(); i++) {
                 const auto& src = resp.items[i];
                 auto item = Napi::Object::New(env);
-                item["status"] = toJs(env, static_cast<uint32_t>(src.status));
+                // status is now an opaque string code (unredeemed/redeemed/expired/revoked; pass through)
+                item["status"] = toJs(env, src.status);
                 item["plan"] = toJs(env, src.plan);
                 item["paymentProvider"] = toJs(env, src.payment_provider);
                 item["autoRenewing"] = toJs(env, src.auto_renewing);
